requirement.txt
Django==1.7.2
Markdown==2.6.2
MySQL-python==1.2.5
Pillow==2.9.0
argparse==1.2.1
django-braces==1.8.1
django-filter==0.10.0
django-oauth-toolkit==0.10.0
django-oauth2-provider==0.2.6.1
django-password-reset==0.7
djangorestframework==3.3.3
ipdb==0.8
ipython==2.2.0
oauthlib==1.0.3
python-dateutil==2.4.2
pytz==2015.4
requests==2.7.0
shortuuid==0.4.3
six==1.9.0
wheel==0.24.0
wsgiref==0.1.2
